separate check_aux
